perf(schema): separate Schema API
